+++ tests/testthat/test-exporters.R
@@ -194,7 +194,39 @@ test_that("as_html smoke test", {
 
     tbl <- tt_to_export()
     oldo <- options(viewer = identity)
-    fl <- Viewer(tbl)
+    expect_silent(fl <- Viewer(tbl))
+    xml2::read_html(fl)
+    expect_true(TRUE)
+    options(oldo)
+})
+
+test_that("as_html Viewer with newline test", {
+
+    tmpf <- tempfile(fileext = ".html")
+    
+    colfuns <- list(function(x) rcell(mean(x), format = "xx.x"),
+                    function(x) rcell(sd(x), format = "xx.x"))
+    varlabs <- c("Mean Age", "SD\nLine Break!!! \nAge")
+    
+    lyt <- basic_table() %>%
+        split_cols_by_multivar(c("AGE", "AGE"), varlabels = varlabs) %>%
+        analyze_colvars(afun = colfuns)
+    
+    tbl_wrapping <- build_table(lyt, DM)
+    
+    tbl_normal <- rtable(
+        header = c("Treatement\nN=100", "Comparison\nN=300"),
+        format = "xx (xx.xx%)",
+        rrow("A", c(104, .2), c(100, .4)),
+        rrow("B", c(23, .4), c(43, .5)),
+        rrow(),
+        rrow("this is a very long section header"),
+        rrow("estimate", rcell(55.23, "xx.xx", colspan = 2)),
+        rrow("95% CI", indent = 1, rcell(c(44.8, 67.4), format = "(xx.x, xx.x)", colspan = 2))
+    )
+    oldo <- options(viewer = identity)
+    expect_silent(fl <- Viewer(tbl_wrapping))
+    expect_silent(fl <- Viewer(tbl_normal))
     xml2::read_html(fl)
     expect_true(TRUE)
     options(oldo)